+++ mods/CC-Tweaked/projects/common/src/main/java/dan200/computercraft/shared/computer/blocks/AbstractComputerBlock.java
@@ -7,6 +7,7 @@ package dan200.computercraft.shared.computer.blocks;
 import dan200.computercraft.annotations.ForgeOverride;
 import dan200.computercraft.shared.common.IBundledRedstoneBlock;
 import dan200.computercraft.shared.network.container.ComputerContainerData;
+import dan200.computercraft.shared.platform.PlatformHelper;
 import dan200.computercraft.shared.platform.RegistryEntry;
 import dan200.computercraft.shared.util.BlockEntityHelpers;
 import net.minecraft.core.BlockPos;
@@ -121,7 +122,7 @@ public abstract class AbstractComputerBlock<T extends AbstractComputerBlockEntit
                 var serverComputer = computer.createServerComputer();
                 serverComputer.turnOn();
 
-                new ComputerContainerData(serverComputer, getItem(computer)).open(player, computer);
+                PlatformHelper.get().openMenu(player, computer.getName(), computer, new ComputerContainerData(serverComputer, getItem(computer)));
             }
             return InteractionResult.sidedSuccess(level.isClientSide);
         }