+++ mods/CC-Tweaked/projects/fabric/src/main/java/dan200/computercraft/shared/platform/PlatformHelperImpl.java
@@ -47,11 +47,11 @@ import net.minecraft.tags.TagKey;
 import net.minecraft.world.Container;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.MenuProvider;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.MenuConstructor;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.*;
 import net.minecraft.world.item.crafting.Ingredient;
@@ -105,8 +105,8 @@ public class PlatformHelperImpl implements PlatformHelper {
     }
 
     @Override
-    public void openMenu(Player player, MenuProvider owner, ContainerData menu) {
-        player.openMenu(new WrappedMenuProvider<>(owner, menu));
+    public void openMenu(Player player, Component title, MenuConstructor menu, ContainerData data) {
+        player.openMenu(new WrappedMenuProvider<>(title, menu, data));
     }
 
     @Override
@@ -277,22 +277,22 @@ public class PlatformHelperImpl implements PlatformHelper {
     }
 
     private record WrappedMenuProvider<T extends ContainerData>(
-        MenuProvider owner, T menu
+        Component title, MenuConstructor menu, T data
     ) implements ExtendedScreenHandlerFactory<T> {
         @Nullable
         @Override
         public AbstractContainerMenu createMenu(int id, Inventory inventory, Player player) {
-            return owner.createMenu(id, inventory, player);
+            return menu.createMenu(id, inventory, player);
         }
 
         @Override
         public Component getDisplayName() {
-            return owner.getDisplayName();
+            return title;
         }
 
         @Override
         public T getScreenOpeningData(ServerPlayer player) {
-            return menu;
+            return data;
         }
     }