+++ mods/CC-Tweaked/projects/core/src/main/resources/data/computercraft/lua/rom/help/changelog.md
@@ -1,3 +1,12 @@
+# New features in CC: Tweaked 1.114.4
+
+* Allow typing/pasting any character in the CC charset.
+* Add a new `computercraft:storage_capacity` to override a disk or computer's capacity.
+
+Several bug fixes:
+* Fix command computers having NBT set when placed in a Create contraption.
+* Use correct bounding box when checking for entities in turtle movement.
+
 # New features in CC: Tweaked 1.114.3
 
 * `wget` now prints the error that occurred, rather than a generic "Failed" (tizu69).